#924 Move plugin_loader.py to the root of neuro-san-studio/plugins
